+++ doers/src/types.rs
@@ -210,8 +210,7 @@ pub struct RemoveResources {
 
 macro_rules! impl_apply {
     ($($t:ty),*) => {
-        $(
-            impl Apply for $t {
+        $( impl Apply for $t {
                 fn id(&self) -> &str { &self.id }
                 fn apply(&self, opts: &ApplyOpts) -> anyhow::Result<ApplySummary> {
                     self.apply(opts)
@@ -274,12 +273,34 @@ pub struct Applicator {
     json_config: String,
 }
 
+#[derive(PartialEq)]
+enum Phase {
+    Check,
+    Apply,
+}
+
 impl Applicator {
     pub fn from(json_config: JsonConfig) -> Self {
         Self { json_config }
     }
 
     pub fn run(&self, opts: &ApplyOpts) -> anyhow::Result<ApplySummary> {
+        if !opts.no_check && !opts.noop {
+            self._run(
+                &ApplyOpts {
+                    noop: true,
+                    ..opts.clone()
+                },
+                Phase::Check,
+            )?;
+
+            tracing::info!("Completed noop check phase");
+        }
+
+        self._run(opts, Phase::Apply)
+    }
+
+    fn _run(&self, opts: &ApplyOpts, phase: Phase) -> anyhow::Result<ApplySummary> {
         let config = self.json_to_hostconfig()?;
         let mut splay = None;
 
@@ -304,17 +325,26 @@ impl Applicator {
             }
         }
 
-        if let Some(max_splay) = splay {
+        if phase == Phase::Apply
+            && let Some(max_splay) = splay
+        {
             let mut rng = rand::rng();
             let time_to_wait = rng.random_range(..=max_splay);
             tracing::info!("splay requested: pausing {time_to_wait} seconds...");
             thread::sleep(Duration::from_secs(time_to_wait));
         }
 
-        tracing::info!("Configuring host: {}", config.metadata.name);
+        match phase {
+            Phase::Check => {
+                tracing::info!("Running noop check phase for: {}", config.metadata.name)
+            }
+            Phase::Apply => tracing::info!("Configuring host: {}", config.metadata.name),
+        }
+
         let ret = self.apply(&config, opts);
 
-        if std::env::var("GURP_RSS_STATS").is_ok()
+        if phase == Phase::Apply
+            && std::env::var("GURP_RSS_STATS").is_ok()
             && let Some(rss) = runtime_stats::rss_bytes()
         {
             tracing::info!("final RSS: {}", ByteSize(rss as u64));